+++ gha_utils/metadata.py
@@ -80,7 +80,7 @@ build_targets=[
     },
     {
         "target": "linux-x64",
-        "os": "ubuntu-slim",
+        "os": "ubuntu-24.04",
         "platform_id": "linux",
         "arch": "x64",
         "extension": "bin"
@@ -117,7 +117,7 @@ build_targets=[
 nuitka_matrix={
     "os": [
         "ubuntu-24.04-arm",
-        "ubuntu-slim",
+        "ubuntu-24.04",
         "macos-26",
         "macos-15-intel",
         "windows-11-arm",
@@ -138,7 +138,7 @@ nuitka_matrix={
         },
         {
             "target": "linux-x64",
-            "os": "ubuntu-slim",
+            "os": "ubuntu-24.04",
             "platform_id": "linux",
             "arch": "x64",
             "extension": "bin"
@@ -201,13 +201,13 @@ nuitka_matrix={
             "bin_name": "mpm-linux-arm64-6f27db4.bin"
         },
         {
-            "os": "ubuntu-slim",
+            "os": "ubuntu-24.04",
             "entry_point": "mpm",
             "commit": "346ce664f055fbd042a25ee0b7e96702e95",
             "bin_name": "mpm-linux-x64-346ce66.bin"
         },
         {
-            "os": "ubuntu-slim",
+            "os": "ubuntu-24.04",
             "entry_point": "mpm",
             "commit": "6f27db47612aaee06fdf08744b09a9f5f6c2",
             "bin_name": "mpm-linux-x64-6f27db4.bin"
@@ -337,7 +337,7 @@ NUITKA_BUILD_TARGETS = {
         "extension": "bin",
     },
     "linux-x64": {
-        "os": "ubuntu-slim",
+        "os": "ubuntu-24.04",
         "platform_id": "linux",
         "arch": "x64",
         "extension": "bin",
@@ -1327,7 +1327,7 @@ class Metadata:
             {
                 "os": [
                     "ubuntu-24.04-arm",
-                    "ubuntu-slim",
+                    "ubuntu-24.04",
                     "macos-26",
                     "macos-15-intel",
                     "windows-11-arm",
@@ -1350,7 +1350,7 @@ class Metadata:
                     },
                     {
                         "target": "linux-x64",
-                        "os": "ubuntu-slim",
+                        "os": "ubuntu-24.04",
                         "platform_id": "linux",
                         "arch": "x64",
                         "extension": "bin",
@@ -1413,13 +1413,13 @@ class Metadata:
                         "bin_name": "mpm-linux-arm64-6f27db4.bin",
                     },
                     {
-                        "os": "ubuntu-slim",
+                        "os": "ubuntu-24.04",
                         "entry_point": "mpm",
                         "commit": "346ce664f055fbd042a25ee0b7e96702e95",
                         "bin_name": "mpm-linux-x64-346ce66.bin",
                     },
                     {
-                        "os": "ubuntu-slim",
+                        "os": "ubuntu-24.04",
                         "entry_point": "mpm",
                         "commit": "6f27db47612aaee06fdf08744b09a9f5f6c2",
                         "bin_name": "mpm-linux-x64-6f27db4.bin",